Fixed ScaleType.CENTER_INSIDE for smaller images
CENTER_INSIDE means the image is (proportionally) scaled
to the shorter dimension of the view. Before that change,
that was only true for images that have a larger length
than one dimension of the view.

Also, there's now ScaleType.CENTER which just centers the
image within view and doesn't allow the image to be scaled
below that initial size.