Remove Travis CI integration
Apparently, at least for Android builds, Travis CI just doesn't
cut the mustard.

After updating the SDK version to 27, Travis CI *always* fails
to build the project because the android-27 license is not accepted.
I've done everything https://docs.travis-ci.com/user/languages/android/
told me to but still, it won't build.

So I've got enough of Travis CI and that's it. If that thing can't
build a standard Android project, it's simply not useful to me.